Separate display of vote counts from voting itself (#8)
Problem
-------

When voting on cards, the total number of votes for each card is shown,
and cards are sorted by number of votes such that the most voted-for
cards are displayed prominently. This may have a psychological effect on
retrospective participants, encouraging them to vote for the most
popular cards rather than cards they would like to see discussed.

Changes
-------

* Display of vote counts is now controlled by a separate menu button in
  moderator view. This state is limited to a single browser and does not
  mutate the session state stored in Redis.